Replace floating types by arbitrary precision decimals (#166)

+++ modules/sql/src/main/scala/Angle.scala
@@ -6,15 +6,15 @@ package gem.sql
 /** Minimal Angle classes to support reading from the enum tables without depending on core. */
 object Angle {
 
-  final case class Arcseconds(toArcsecs: Double)
+  final case class Arcseconds(toArcsecs: BigDecimal)
   object Arcseconds {
-    def fromArcsecs(d: Double): Arcseconds =
+    def fromArcsecs(d: BigDecimal): Arcseconds =
       Arcseconds(d)
   }
 
-  final case class Degrees(toDegrees: Double)
+  final case class Degrees(toDegrees: BigDecimal)
   object Degrees {
-    def fromDegrees(d: Double): Degrees =
+    def fromDegrees(d: BigDecimal): Degrees =
       Degrees(d)
   }
